experiments/runner/pyproject.toml
[project]
name = "icl-alignment-experiments"
version = "0.1.0"
description = "Experiments exploring emergent alignment via in-context learning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.55.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "pandas>=2.2.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.scripts]
icl = "icl_experiments.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/icl_experiments"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ipython>=8.0.0",
]
